Team_MemManage/NcertRAGChatbot_Project/requirements.txt
llama-index==0.10.30
google-generativeai
deepeval
PyMuPDF
tiktoken
llama-index-retrievers-bm25
streamlit